apply single reposnabiliry at content folder
