+++ frontend/.gitignore
@@ -5,7 +5,7 @@
 /node_modules
 /.pnp
 .pnp.js
-
+claude.md
 # testing
 /coverage
