+++ frontend/src/locales/en/messages.json
@@ -1,5 +1,5 @@
 {
-  "app_title": {"message": "VIX Accessibility Assistant"},
+  "app_title": {"message": "VIX Accessibility Assistant 2"},
   "app_description": {"message": "Intelligent system to improve web accessibility"},
   "connecting": {"message": "Connecting..."},
   "dom_status": {"message": "DOM Status"},
